requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.6
django-autocomplete-light==3.11.0
django-classy-tags==4.1.0
django-cms==4.1.1
django-cors-headers==4.4.0
django-filter==24.2
django-formtools==2.5.1
django-haystack==3.3.0
django-modelcluster==6.3
django-mysql==4.14.0
django-permissionedforms==0.1
django-sekizai==4.1.0
django-taggit==5.0.1
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-text-ckeditor==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
draftjs-exporter==5.0.0
elastic-transport==8.13.1
elasticsearch==8.14.0
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.7
l18n==2021.3
laces==0.1.1
mysqlclient==2.2.4
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pillow_heif==0.17.0
PyJWT==2.8.0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
telepath==0.3.1
typing_extensions==4.12.2
urllib3==2.2.2
wagtail==6.1.2
webencodings==0.5.1
Willow==1.8.0
